+++ decidim-core/app/controllers/decidim/devise/confirmations_controller.rb
@@ -6,6 +6,8 @@ module Decidim
     class ConfirmationsController < ::Devise::ConfirmationsController
       include Decidim::DeviseControllers
 
+      helper_method :new_user_group_session_path
+
       # Since we're using a single Devise installation for multiple
       # organizations, and user emails can be repeated across organizations,
       # we need to identify the user by both the email and the organization.
@@ -17,6 +19,13 @@ module Decidim
       # `decidim_organization_id` attribute.
       def resource_params
         super.merge(decidim_organization_id: current_organization.id)
+      end
+
+      # Overwrites the default method to handle user groups confirmations.
+      def after_confirmation_path_for(resource_name, resource)
+        return profile_path(resource.nickname) if resource_name == :user_group
+
+        super
       end
     end
   end