+++ CHANGELOG.md
@@ -96,6 +96,7 @@ In order to generate Open Data exports you should add this to your crontab or re
 - **decidim-core**: Adds a user timeline tab on the public profile. [\#4574](https://github.com/decidim/decidim/pull/4574)
 - **decidim-core**: Open Data export [\#4578](https://github.com/decidim/decidim/pull/4578)
 - **decidim-meetings**: Export meetings [\#4597](https://github.com/decidim/decidim/pull/4597)
+- **decidim-core**: User groups can now confirm their email [\#4603](https://github.com/decidim/decidim/pull/4603)
 
 **Changed**: